+++ metacubexd-main/src/pages/Config.tsx
@@ -1,6 +1,6 @@
 import { createForm } from '@felte/solid'
 import { validator } from '@felte/validator-zod'
-import type { Accessor, Component, JSX, ParentComponent } from 'solid-js'
+import type { Accessor, JSX, ParentComponent } from 'solid-js'
 import { toast } from 'solid-toast'
 import { twMerge } from 'tailwind-merge'
 import { z } from 'zod'
@@ -9,8 +9,6 @@ import {
   fetchBackendVersionAPI,
   flushFakeIPDataAPI,
   flushingFakeIPData,
-  isBackendUpdateAvailableAPI,
-  isFrontendUpdateAvailableAPI,
   reloadConfigFileAPI,
   reloadingConfigFile,
   restartBackendAPI,
@@ -18,12 +16,8 @@ import {
   updateBackendConfigAPI,
   updateGEODatabasesAPI,
   updatingGEODatabases,
-  upgradeBackendAPI,
-  upgradeUIAPI,
-  upgradingBackend,
-  upgradingUI,
 } from '~/apis'
-import { Button, ConfigTitle, DocumentTitle } from '~/components'
+import { Button, ConfigTitle, DocumentTitle, Versions } from '~/components'
 import { LANG, ROUTES, themes } from '~/constants'
 import { Dict, locale, setLocale, useI18n } from '~/i18n'
 import {
@@ -56,12 +50,7 @@ const Input: ParentComponent<JSX.InputHTMLAttributes<HTMLInputElement>> = (
 ) => {
   const [local, others] = splitProps(props, ['class'])
 
-  return (
-    <input
-      class={twMerge('input input-bordered min-w-0', local.class)}
-      {...others}
-    />
-  )
+  return <input class={twMerge('input min-w-0', local.class)} {...others} />
 }
 
 const Select: ParentComponent<JSX.SelectHTMLAttributes<HTMLSelectElement>> = (
@@ -70,7 +59,7 @@ const Select: ParentComponent<JSX.SelectHTMLAttributes<HTMLSelectElement>> = (
   const [local, others] = splitProps(props, ['class'])
 
   return (
-    <select class={twMerge('select select-bordered', local.class)} {...others}>
+    <select class={twMerge('select', local.class)} {...others}>
       {children(() => others.children)()}
     </select>
   )
@@ -82,10 +71,8 @@ const Label: ParentComponent<JSX.LabelHTMLAttributes<HTMLLabelElement>> = (
   const [local, others] = splitProps(props, ['class'])
 
   return (
-    <label class={twMerge('label', local.class)} {...others}>
-      <span class="label-text truncate">
-        {children(() => others.children)()}
-      </span>
+    <label class={twMerge('label truncate', local.class)} {...others}>
+      {children(() => others.children)()}
     </label>
   )
 }
@@ -98,6 +85,7 @@ const dnsQueryFormSchema = z.object({
 const DNSQueryForm = () => {
   const [t] = useI18n()
   const request = useRequest()
+  const defaultDNSQueryTarget = 'google.com'
 
   const { form, isSubmitting } = createForm<z.infer<typeof dnsQueryFormSchema>>(
     {
@@ -105,7 +93,10 @@ const DNSQueryForm = () => {
       onSubmit: (values) =>
         request
           .get('dns/query', {
-            searchParams: { name: values.name, type: values.type },
+            searchParams: {
+              name: values.name || defaultDNSQueryTarget,
+              type: values.type,
+            },
           })
           .json<DNSQuery>()
           .then(({ Answer }) =>
@@ -124,7 +115,7 @@ const DNSQueryForm = () => {
           type="search"
           name="name"
           class="flex-1"
-          placeholder="google.com"
+          placeholder={defaultDNSQueryTarget}
           onInput={(e) => {
             if (!e.target.value) setDNSQueryResult([])
           }}
@@ -164,8 +155,7 @@ const configFormSchema = z.object({
 
 const ConfigForm: ParentComponent<{
   isSingBox: Accessor<boolean>
-  fetchBackendVersion: () => Promise<void>
-}> = ({ isSingBox, fetchBackendVersion }) => {
+}> = ({ isSingBox }) => {
   const [t] = useI18n()
 
   const portList = [
@@ -241,7 +231,7 @@ const ConfigForm: ParentComponent<{
   return (
     <div class="flex flex-col gap-4">
       <div class="grid grid-cols-3 gap-2">
-        <div class="form-control">
+        <fieldset class="fieldset">
           <Label for="enable-allow-lan">{t('allowLan')}</Label>
 
           <Toggle
@@ -255,9 +245,9 @@ const ConfigForm: ParentComponent<{
               )
             }
           />
-        </div>
+        </fieldset>
 
-        <div class="form-control">
+        <fieldset class="fieldset">
           <Label for="mode">{t('runningMode')}</Label>
 
           <Select
@@ -274,9 +264,9 @@ const ConfigForm: ParentComponent<{
               )}
             </For>
           </Select>
-        </div>
+        </fieldset>
 
-        <div class="form-control">
+        <fieldset class="fieldset">
           <Label for="interface-name">{t('outboundInterfaceName')}</Label>
 
           <Input
@@ -290,12 +280,12 @@ const ConfigForm: ParentComponent<{
               )
             }
           />
-        </div>
+        </fieldset>
       </div>
 
       <Show when={!isSingBox()}>
         <div class="grid grid-cols-3 gap-2">
-          <div class="form-control">
+          <fieldset class="fieldset">
             <Label for="enable-tun-device">{t('enableTunDevice')}</Label>
 
             <Toggle
@@ -309,9 +299,9 @@ const ConfigForm: ParentComponent<{
                 )
               }
             />
-          </div>
+          </fieldset>
 
-          <div class="form-control">
+          <fieldset class="fieldset">
             <Label for="tun-ip-stack">{t('tunModeStack')}</Label>
 
             <Select
@@ -335,9 +325,9 @@ const ConfigForm: ParentComponent<{
                 )}
               </For>
             </Select>
-          </div>
+          </fieldset>
 
-          <div class="form-control">
+          <fieldset class="fieldset">
             <Label for="device-name">{t('tunDeviceName')}</Label>
 
             <Input
@@ -351,13 +341,13 @@ const ConfigForm: ParentComponent<{
                 )
               }
             />
-          </div>
+          </fieldset>
         </div>
 
         <form class="grid grid-cols-3 gap-2 sm:grid-cols-5" use:form={form}>
           <For each={portList}>
             {(item) => (
-              <div class="form-control">
+              <fieldset class="fieldset">
                 <Label for={item.key}>{item.label()}</Label>
 
                 <Input
@@ -367,13 +357,13 @@ const ConfigForm: ParentComponent<{
                   placeholder={item.label()}
                   onChange={item.onChange}
                 />
-              </div>
+              </fieldset>
             )}
           </For>
         </form>
       </Show>
 
-      <div class="grid grid-cols-2 gap-4 sm:grid-cols-3">
+      <div class="grid grid-cols-2 gap-4 sm:grid-cols-2">
         <Button
           class="btn-primary"
           loading={reloadingConfigFile()}
@@ -406,25 +396,6 @@ const ConfigForm: ParentComponent<{
           >
             {t('updateGEODatabases')}
           </Button>
-
-          <Button
-            class="btn-info"
-            loading={upgradingUI()}
-            onClick={upgradeUIAPI}
-          >
-            {t('upgradeUI')}
-          </Button>
-
-          <Button
-            class="btn-error"
-            loading={upgradingBackend()}
-            onClick={async () => {
-              await upgradeBackendAPI()
-              await fetchBackendVersion()
-            }}
-          >
-            {t('upgradeCore')}
-          </Button>
         </Show>
       </div>
     </div>
@@ -442,6 +413,10 @@ const ConfigForXd = () => {
     {
       label: () => t('zh'),
       value: LANG.ZH,
+    },
+    {
+      label: () => t('ru'),
+      value: LANG.RU,
     },
   ]
 
@@ -473,7 +448,7 @@ const ConfigForXd = () => {
 
         <div class="flex flex-col">
           <ConfigTitle>{endpoint()?.url}</ConfigTitle>
-          
+
           <Button
             class="btn-info"
             onClick={() => {
@@ -540,48 +515,6 @@ const ConfigForXd = () => {
   )
 }
 
-const Versions: Component<{
-  frontendVersion: string
-  backendVersion: Accessor<string>
-}> = ({ frontendVersion, backendVersion }) => {
-  const [isFrontendUpdateAvailable] = createResource(() =>
-    isFrontendUpdateAvailableAPI(frontendVersion),
-  )
-  const [isBackendUpdateAvailable, { refetch: fetchIsBackendUpdateAvailable }] =
-    createResource(() => isBackendUpdateAvailableAPI(backendVersion()))
-
-  createEffect(() => {
-    fetchIsBackendUpdateAvailable()
-  }, backendVersion())
-
-  const UpdateAvailableIndicator = () => (
-    <span class="absolute -right-1 -top-1 flex h-3 w-3">
-      <span class="absolute inline-flex h-full w-full animate-ping rounded-full bg-info opacity-75" />
-      <span class="inline-flex h-3 w-3 rounded-full bg-info" />
-    </span>
-  )
-
-  return (
-    <div class="grid grid-cols-2 gap-4 mx-2 md:mx-0">
-      <div class="relative">
-        <Show when={isFrontendUpdateAvailable()}>
-          <UpdateAvailableIndicator />
-        </Show>
-
-        <kbd class="kbd w-full">{import.meta.env.APP_VERSION}</kbd>
-      </div>
-
-      <div class="relative">
-        <Show when={isBackendUpdateAvailable()}>
-          <UpdateAvailableIndicator />
-        </Show>
-
-        <kbd class="kbd w-full">{backendVersion()}</kbd>
-      </div>
-    </div>
-  )
-}
-
 export default () => {
   const navigate = useNavigate()
 
@@ -594,10 +527,9 @@ export default () => {
   const [t] = useI18n()
 
   const frontendVersion = `v${import.meta.env.APP_VERSION}`
-  const [backendVersion, { refetch: fetchBackendVersion }] = createResource(
-    fetchBackendVersionAPI,
-    { initialValue: '' },
-  )
+  const [backendVersion] = createResource(fetchBackendVersionAPI, {
+    initialValue: '',
+  })
 
   const isSingBox = createMemo(
     () => backendVersion()?.includes('sing-box') || false,
@@ -616,12 +548,7 @@ export default () => {
 
         <ConfigTitle withDivider>{t('coreConfig')}</ConfigTitle>
 
-        <ConfigForm
-          isSingBox={isSingBox}
-          fetchBackendVersion={() =>
-            fetchBackendVersion() as unknown as Promise<void>
-          }
-        />
+        <ConfigForm isSingBox={isSingBox} />
 
         <ConfigTitle withDivider>{t('xdConfig')}</ConfigTitle>